+++ hw2-configuration-spring-app/src/main/java/ru/quizapp/service/ExamTicketService.java
@@ -1,9 +1,15 @@
 package ru.quizapp.service;
 
 import ru.quizapp.dto.ExamTicketDTO;
+import ru.quizapp.dto.ExaminationDTO;
+import ru.quizapp.dto.StudentDTO;
 
 import java.util.List;
 
 public interface ExamTicketService {
     List<ExamTicketDTO> getAllTickets();
+
+    StudentDTO studentRegistration();
+
+    ExaminationDTO studentTesting(StudentDTO student);
 }
